requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
audiocraft==0.0.2a2
audioread==3.0.0
av==10.0.0
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorlog==6.7.0
confection==0.0.4
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.35
decorator==5.1.1
demucs==4.0.0
diffq==0.2.4
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
docopt==0.6.2
dora-search==0.1.12
einops==0.6.1
fastapi==0.97.0
ffmpy==0.3.0
filelock==3.12.2
flashy==0.0.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gradio==3.35.2
gradio_client==0.2.7
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
julius==0.2.7
kiwisolver==1.4.4
lameenc==1.4.2
langcodes==3.3.0
lazy_loader==0.2
librosa==0.10.0.post2
linkify-it-py==2.0.2
llvmlite==0.40.1rc1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
num2words==0.5.12
numba==0.57.0
numpy==1.24.3
omegaconf==2.3.0
openai==0.27.8
openunmix==1.2.1
orjson==3.9.1
packaging==23.1
pandas==2.0.2
pathy==0.10.1
Pillow==9.5.0
pooch==1.6.0
preshed==3.0.8
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
retrying==1.3.4
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.5
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.6
starlette==0.27.0
submitit==1.4.5
sympy==1.12
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
tqdm==4.65.0
transformers==4.30.2
treetable==0.2.5
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.3
uvicorn==0.22.0
wasabi==1.1.2
websockets==11.0.3
xformers==0.0.20
yarl==1.9.2
